Issue #214: Moved Settings modules into own components
